Ner_Pipeline/requirements.txt
absl-py==2.3.1
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
ale-py==0.11.2
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-generator==1.10
async-lru==2.0.5
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
audioread==3.0.1
autorom==0.6.1
autorom-accept-rom-license==0.6.1
azure-common==1.1.28
azure-core==1.35.0
azure-mgmt-core==1.6.0
azure-mgmt-resource==24.0.0
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
bioc==2.1
bitsandbytes==0.46.1
bleach==6.2.0
blis==0.7.11
boto3==1.35.99
botocore==1.35.99
build==1.2.2.post1
cachetools==6.1.0
catalogue==2.0.10
celery==5.4.0
certifi==2025.7.14
certipy==0.2.2
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-params==0.4.0
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.21.1
cloudpickle==3.1.1
colorama==0.4.6
colorcet==3.1.0
coloredlogs==15.0.1
colorspacious==1.1.2
comm==0.2.2
conda-store==2025.2.2
confection==0.1.5
configobj==5.0.9
conllu==4.5.3
contourpy==1.3.2
cryptography==43.0.3
cuda-bindings==12.9.0
cuda-python==12.9.0
cycler==0.12.1
cymem==2.0.11
datasets==3.0.2
datashader==0.18.1
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
deprecated==1.2.18
dictdiffer==0.9.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docker-pycreds==0.4.0
docopt==0.6.2
docstring-parser==0.16
dpath==2.2.0
dulwich==0.23.2
dvc==3.55.2
dvc-data==3.16.10
dvc-http==2.32.0
dvc-objects==5.1.1
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
en-core-sci-md @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_md-0.5.4.tar.gz
entrypoints==0.4
evaluate==0.4.5
executing==2.2.0
farama-notifications==0.0.4
fastapi==0.116.1
fastjsonschema==2.21.1
filelock==3.18.0
flair==0.15.0
flatbuffers==25.2.10
flatten-dict==0.4.2
flufl-lock==8.2.0
fonttools==4.59.0
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2024.9.0
ftfy==6.3.1
funcy==2.0
gdown==5.2.0
gitdb==4.0.12
gitpython==3.1.44
grandalf==0.8
greenlet==3.2.3
grpcio==1.73.1
gto==1.7.2
gymnasium==0.29.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.24.5
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
importlib-metadata==8.7.0
importlib-resources==6.5.2
iniconfig==2.1.0
intervaltree==3.1.0
ipinfo==5.1.1
ipykernel==6.29.5
ipython==9.4.0
ipython-genutils==0.2.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.7
isodate==0.7.2
isoduration==20.11.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
json5==0.12.0
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-core==5.8.1
jupyter-events==0.12.0
jupyter-highlight-selected-word==0.2.0
jupyter-launcher-shortcuts==4.0.3
jupyter-lsp==2.2.5
jupyter-nbextensions-configurator==0.6.4
jupyter-server==2.16.0
jupyter-server-terminals==0.5.3
jupyter-telemetry==0.1.0
jupyterhub==5.3.0
jupyterlab==4.4.9
jupyterlab-conda-store==2024.11.1
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.15
kiwisolver==1.4.8
kombu==5.4.2
lab==8.4
langcodes==3.5.0
langdetect==1.0.9
language-data==1.3.0
lazy-loader==0.4
levenshtein==0.26.1
llvmlite==0.43.0
locket==1.0.0
loguru==0.7.3
lxml==6.0.0
mako==1.3.10
marisa-trie==1.2.1
markdown==3.8.2
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
more-itertools==10.7.0
mpld3==0.5.10
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multipledispatch==1.0.0
multiprocess==0.70.16
murmurhash==1.0.13
nb-conda-store-kernels==0.1.5
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nervaluate==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nmslib-metabrainz==2.1.3
nodejs==0.1.1
notebook==7.4.7
notebook-shim==0.2.4
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.16.2
onnxruntime==1.19.0
opencv-python==4.10.0.84
optimum==1.26.1
optional-django==0.1.0
optuna==4.5.0
orjson==3.10.7
overrides==7.7.0
packaging==25.0
pamela==1.2.0
pandas==2.2.2
pandocfilters==1.5.1
param==2.2.0
parso==0.8.4
partd==1.4.2
passlib==1.7.4
pathlib-abc==0.1.1
pathspec==0.12.1
pathy==0.11.0
peft==0.12.0
pexpect==4.9.0
pillow==10.4.0
pip==23.1.2
pip-tools==7.4.1
platformdirs==3.11.0
plotly==5.24.1
pluggy==1.6.0
pooch==1.8.2
pptree==3.1
preshed==3.0.10
prometheus-client==0.20.0
prompt-toolkit==3.0.51
protobuf==3.20.2
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==17.0.0
pyarrow-hotfix==0.6
pybind11==2.13.6
pycparser==2.22
pyct==0.5.0
pydantic==2.8.2
pydantic-core==2.20.1
pydantic-settings==2.9.1
pydot==3.0.1
pygame==2.6.0
pygit2==1.15.1
pygments==2.19.1
pygtrie==2.5.0
pyjwt==2.7.0
pylabeladjust==0.1.13
pymysql==1.1.1
pynndescent==0.5.13
pyopenssl==24.2.1
pyparsing==3.2.3
pyproject-hooks==1.2.0
pyqtree==1.0.0
pysbd==0.3.4
pysocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==2.0.7
python-levenshtein==0.26.1
python-multipart==0.0.20
pytorch-revgrad==0.2.0
pytz==2024.1
pyyaml==6.0.2
pyzmq==26.1.0
qtpy==2.4.1
rapidfuzz==3.12.1
rcssmin==1.2.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-file==2.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rjsmin==1.2.3
rpds-py==0.20.0
ruamel-yaml==0.18.6
ruamel-yaml-clib==0.2.8
s3transfer==0.10.4
safetensors==0.4.4
scikit-image==0.25.0
scikit-learn==1.5.1
scipy==1.14.0
scispacy==0.5.5
scmrepo==3.3.7
seaborn==0.13.2
secure==0.3.0
segtok==1.5.11
semver==3.0.2
send2trash==1.8.3
sentence-transformers==3.3.1
sentencepiece==0.2.0
sentry-sdk==2.14.0
seqeval==1.2.2
setproctitle==1.3.3
setuptools==65.5.0
shellingham==1.5.4
shimmy==1.3.0
shortuuid==1.0.13
shtab==1.7.1
simplejson==3.19.2
six==1.17.0
smart-open==6.4.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
sox==1.5.0
soxr==0.5.0.post1
spacy==3.7.5
spacy-alignments==0.9.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.9
sqlalchemy==2.0.41
sqlalchemy-utils==0.38.3
sqlalchemy2-stubs==0.0.2a38
sqlitedict==2.1.0
sqlmodel==0.0.18
sqltrie==0.11.1
srsly==2.5.1
stable-baselines3==2.3.2
stack-data==0.6.3
starlette==0.45.3
sympy==1.13.2
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
terminado==0.18.1
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2024.12.12
tinycss2==1.2.1
tldextract==5.1.3
tokenizers==0.21.0
tomlkit==0.13.2
toolz==1.0.0
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformer-smaller-training-vocab==0.4.0
transformers==4.48.2
triton==3.0.0
trl==0.9.6
txt2tags==3.9
typer==0.16.0
types-python-dateutil==2.9.0.20240316
typing-extensions==4.13.2
typing-inspection==0.4.1
tyro==0.8.6
tzdata==2024.1
uri-template==1.3.0
urllib3==2.4.0
uv==0.7.5
uvicorn==0.34.2
uvloop==0.21.0
validators==0.18.2
vine==5.1.0
voluptuous==0.15.2
wandb==0.17.9
wasabi==1.1.3
watchfiles==1.0.5
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
werkzeug==3.0.4
wheel==0.45.1
widgetsnbextension==4.0.14
wikipedia-api==0.7.1
wrapt==1.17.2
xarray==2024.11.0
xlrd==2.0.1
xxhash==3.4.1
yarl==1.9.4
zc-lockfile==3.0.post1
zipp==3.21.0
